Sprint 8.5, filtros e busca finalizada

+++ catalog-api/src/main/java/br/com/catalog/api/dto/ProdutoResponse.java
@@ -34,7 +34,7 @@ public class ProdutoResponse {
     private String artesaoNomeAtelie;
     private Long artesaoId;
     private Boolean artesaoSeloVerificado;
-    private Double notaMedia;
+    private BigDecimal notaMedia;
     private Integer totalAvaliacoes;
     private List<String> imagensUrls;
     private LocalDateTime criadoEm;
